requirements.txt
streamlit==1.31.0
boto3==1.34.22
python-dotenv==1.0.0 